+++ packages/core/src/components/ozon-content/functions/input-to-xml-document.function.ts
@@ -1,0 +1,9 @@
+import { OzonContentInputType } from "../ozon-content.interfaces";
+
+export function inputToXmlDocument(input: OzonContentInputType, domParser: DOMParser): XMLDocument {
+  if (input instanceof XMLDocument) {
+    return input;
+  }
+
+  return domParser.parseFromString(input, "text/xml");
+}